Add files via upload
This simple embedded systems project demonstrates LED blinking using the 8051 (80C51) microcontroller. The LED is connected to Port 2, Pin 0 (P2.0), and it toggles ON and OFF with a delay of 50 milliseconds, creating a visible blinking effect.
The project is written in C language using the Keil uVision IDE and simulates delay using nested for loops. This is a beginner-level project to understand basic GPIO operations and timing with the 8051 microcontroller.

Features :

Microcontroller: 80C51 / AT89C51
Programming Language: C
LED connected to P2.0
Delay-based toggling to create a blinking effect